requirements.txt
aiofiles==0.7.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
ecdsa==0.14.1
fastapi==0.65.1
h11==0.12.0
idna==2.10
motor==2.3.1
odmantic==0.3.5
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.8.2
pymongo==3.11.4
python-dotenv==0.17.1
python-jose==3.2.0
python-multipart==0.0.5
python-slugify==5.0.2
requests==2.25.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
text-unidecode==1.3
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.13.4
